requirements.txt
aiohttp
beautifulsoup4
discord.py
python-dotenv